fix(condense): use FLOAT16 instead of BFLOAT16 for percentage fields (Bug 310)
BFLOAT16 (7 mantissa bits) caused ±0.12% rounding errors for @Percentage fields
(e.g. G1AdaptiveIHOP.thresholdPercentage: 47.37% → 47.46%). FLOAT16 (IEEE 754
half-precision, 10 mantissa bits) uses the same 16-bit storage but reduces the
error to ±0.025%, matching the oracle's displayed values.

Existing CJFR files still decode correctly since the type tag is stored in-file.

+++ src/main/java/me/bechberger/jfr/BasicJFRWriter.java
@@ -762,7 +762,7 @@ public class BasicJFRWriter {
     private FloatType getPercentageFloatType() {
         if (percentageFloatType == null) {
             percentageFloatType =
-                    out.writeAndStoreType(id -> new FloatType(id, "percentage", "", Type.BFLOAT16));
+                    out.writeAndStoreType(id -> new FloatType(id, "percentage", "", Type.FLOAT16));
         }
         return percentageFloatType;
     }
